Add files via upload
GitHub required a PDF of my project to be uploaded, therefore the video in the PDF (just after the slide titled "in the following demonstration") will not play. There will be a secondary attachment in this repository containing the video, or a link to a YouTube upload if there are further complications.